progress on home_screen
