+++ lib/l10n/app_en.arb
@@ -366,5 +366,8 @@
     "moreResources": "More resources",
     "trainingMentoringFirstSteps": "Training: Mentoring First Steps",
     "trainingBuildingFuture": "Training: Building a Sustainable Future",
-    "trainingMarketingBusiness": "Marketing Your Business 101"
+    "trainingMarketingBusiness": "Marketing Your Business 101",
+    "mentorsForYou": "Mentors for you",
+    "recommendationsBasedOnYourProfile": "Recommendations are based on profile",
+    "seeMore": "See more"
 }
